refactor(forma/plot): extract AxisRange into its own header
AxisRange is a standalone type used by geometry functions, SeriesBuilder,
and user raw lambdas. Keeping it in Plot.hpp forced a full geometry function
include for a type dependency. Now in AxisRange.hpp with no further includes
beyond std::function.

apply_auto_scale updated to check auto_scaling field and evaluate
scale_predicate before applying the data-derived range.
